update utils package to use hcam_widgets
