+++ hcam_drivers/utils/__init__.py
@@ -9,6 +9,5 @@
 
 from __future__ import print_function, unicode_literals, absolute_import, division
 
-
-class DriverError(Exception):
-    pass
+__author__ = """Stuart Littlefair"""
+__email__ = 's.littlefair@shef.ac.uk'
